api: split openapi error examples
