+++ quality/quality_scorecard.md
@@ -11,7 +11,7 @@ evidence for PR readiness, not a completion claim.
 | Largest behavior units | Largest function/class included `calculate_risk` at 284 lines, `calculate_rolling_metrics` at 230 lines, and `LotusPerformanceClient` at 256 lines | Largest remaining function is `_issuer_concentration` at 37 lines; `LotusPerformanceClient` is 52 lines | Large engines and clients were decomposed into helpers, services, routers, and polling/parsing functions | Continue reducing engine-level orchestration hotspots |
 | Complexity | Baseline reported C-or-worse candidates across large service, contract, and readiness code | Current baseline reports no C-or-worse candidates in the complexity snapshot | C-level candidates in concentration parsing, risk period resolution, rolling/attribution validation, and enterprise authorization were removed | Keep radon report-only evidence clean while thresholds are tightened |
 | Architecture enforcement | Import-linter, architecture docs, and quality workflow were introduced as report-only baseline | `make architecture-gate` is green locally and in feature-lane CI | Architecture boundary checks are now part of routine slice validation | Extend contracts as service boundaries mature |
-| OpenAPI governance | Operation IDs were not visibly standardized; route-level examples needed certification after router extraction | Operation IDs are explicit; JSON mutation request examples are modularized and enforced by `make openapi-gate`; standard error examples are builder-backed and include additive RFC 7807/problem-details fields; generated artifact policy is enforced by `make openapi-artifact-gate`; current artifact checksum evidence is recorded in `quality/openapi_artifact_evidence.md` | OpenAPI metadata is easier to review, no longer buried in large contract classes, and now fails missing operation IDs/request examples and missing generated artifact evidence in CI lanes while preserving the Lotus error envelope | Regenerate and attach the final current OpenAPI artifact in the PR |
+| OpenAPI governance | Operation IDs were not visibly standardized; route-level examples needed certification after router extraction | Operation IDs are explicit; JSON mutation request examples and standard error examples are modularized and enforced by `make openapi-gate`; standard error examples are builder-backed and include additive RFC 7807/problem-details fields; generated artifact policy is enforced by `make openapi-artifact-gate`; current artifact checksum evidence is recorded in `quality/openapi_artifact_evidence.md` | OpenAPI metadata is easier to review, no longer buried in large contract classes or runtime exception-handler code, and now fails missing operation IDs/request examples and missing generated artifact evidence in CI lanes while preserving the Lotus error envelope | Regenerate and attach the final current OpenAPI artifact in the PR |
 | Tests | 77 Python test files at initial baseline; repo-native coverage gate existed | 103 Python test files; 551 tests collected in the latest baseline; OpenAPI gate logic has focused regression tests | Focused unit/integration coverage protects router, client, contract, middleware, service, and OpenAPI-governance refactors | Add more negative/security contract certification tests |
 | Security | Enterprise audit middleware, redaction tests, and upstream error mapping existed; abuse-control evidence was still a gap | Authorization checks, enterprise audit/redaction, and policy metadata are decomposed into dedicated modules; enterprise runtime and unmapped writes fail closed; correlation/trace input, downstream errors, response headers, and downstream base URLs are hardened with negative tests; threat-model and deployment policy evidence is pinned; Bandit and pip-audit remain green in baseline | Security behavior, deployment posture, unsafe-input handling, audit metadata handling, and abuse controls are easier to inspect and test without changing valid local-development semantics | Add gateway-backed token-validation evidence and final target-runtime configuration proof before release promotion |
 | Observability | HTTP, endpoint execution, supportability, freshness metrics, and correlation existed but needed consolidated docs | Observability docs, dashboard panels, alert definitions, runbook anchors, and endpoint/upstream metrics are covered by tests and baseline validation | Metrics/correlation posture is preserved through router and client decomposition, and operator response evidence is now governed | Keep alert thresholds aligned with production telemetry after deployment |
